INCOMPLETE this is incomplete work from Feb 2016
This is part of a partially working plugin for GDB to allow
breakpoints and stack unwinding in the generated Psyco JIT code,
in theory you should be able to break on Python functions and code
locations etc, but it doesn't work properly in most cases.
it requires the patched fork of GDB on my GitHub.
I think I was trying to understand how to deal with inlined functions.

+++ c/codemanager.c
@@ -224,6 +224,7 @@ DEFINEVAR CodeBufferObject* psyco_codebuf_chained_list = NULL;
 DEFINEVAR void** psyco_codebuf_spec_dict_list = NULL;
 #endif
 
+
 DEFINEFN
 void psyco_shrink_code_buffer(CodeBufferObject* obj, code_t* codeend)
 {
@@ -232,7 +233,8 @@ void psyco_shrink_code_buffer(CodeBufferObject* obj, code_t* codeend)
   debug_printf(3, ("disassemble %p %p    (%d bytes)\n", obj->codestart,
                    codeend, codeend - ((code_t*)obj->codestart)));
   if (VERBOSE_LEVEL == 2)
-    fprintf(stderr, "[%ld]", codeend - ((code_t*)obj->codestart));
+    fprintf(stderr, "[%ld]", (long)(codeend - 
+            ((code_t*)obj->codestart)));
   
   close_buffer_use(codeend);
   SET_CODEMODE(obj, "normal");
